+++ packages/devkit/internal-testing-utils.ts
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/no-restricted-imports */
-export * from 'nx/internal-testing-utils/assert-valid-migrations';
-export * from 'nx/internal-testing-utils/run-migration-against-this-workspace';
-export * from 'nx/internal-testing-utils/with-environment';
+export * from 'nx/src/internal-testing-utils/assert-valid-migrations';
+export * from 'nx/src/internal-testing-utils/run-migration-against-this-workspace';
+export * from 'nx/src/internal-testing-utils/with-environment';
+export * from 'nx/src/internal-testing-utils/temp-fs';
