+++ packages/eslint-plugin/src/rules/dependency-checks.spec.ts
@@ -1,4 +1,4 @@
-import 'nx/src/utils/testing/mock-fs';
+import 'nx/src/internal-testing-utils/mock-fs';
 
 import dependencyChecks, {
   Options,
