+++ packages/eslint-plugin/src/utils/runtime-lint-utils.spec.ts
@@ -1,4 +1,4 @@
-import 'nx/src/utils/testing/mock-fs';
+import 'nx/src/internal-testing-utils/mock-fs';
 
 import {
   ProjectGraph,
