+++ packages/js/src/utils/package-json/update-package-json.spec.ts
@@ -1,4 +1,4 @@
-import 'nx/src/utils/testing/mock-fs';
+import 'nx/src/internal-testing-utils/mock-fs';
 
 import {
   getUpdatedPackageJsonContent,
