+++ packages/nx/migrations.spec.ts
@@ -1,6 +1,6 @@
 import json = require('./migrations.json');
 
-import { assertValidMigrationPaths } from './internal-testing-utils/assert-valid-migrations';
+import { assertValidMigrationPaths } from './src/internal-testing-utils/assert-valid-migrations';
 import { MigrationsJson } from './src/config/misc-interfaces';
 
 describe('nx migrations', () => {
