+++ packages/nx/src/internal-testing-utils/mock-fs.ts
