+++ packages/nx/src/internal-testing-utils/temp-fs.ts
@@ -10,8 +10,9 @@ import {
   rmSync,
   unlinkSync,
 } from 'fs-extra';
-import { joinPathFragments } from '../path';
+import { joinPathFragments } from '../utils/path';
 import { appendFileSync, existsSync, renameSync, writeFileSync } from 'fs';
+import { setWorkspaceRoot, workspaceRoot } from '../utils/workspace-root';
 
 type NestedFiles = {
   [fileName: string]: string;
@@ -19,10 +20,19 @@ type NestedFiles = {
 
 export class TempFs {
   readonly tempDir: string;
+
+  private previousEnvWorkspaceRootPath: string;
+  private previousWorkspaceRoot: string;
+
   constructor(private dirname: string, overrideWorkspaceRoot = true) {
     this.tempDir = realpathSync(mkdtempSync(join(tmpdir(), this.dirname)));
+
+    this.previousEnvWorkspaceRootPath = process.env.NX_WORKSPACE_ROOT_PATH;
+    this.previousWorkspaceRoot = workspaceRoot;
+
     if (overrideWorkspaceRoot) {
       process.env.NX_WORKSPACE_ROOT_PATH = this.tempDir;
+      setWorkspaceRoot(this.tempDir);
     }
   }
 
@@ -76,6 +86,8 @@ export class TempFs {
 
   cleanup() {
     rmSync(this.tempDir, { recursive: true });
+    process.env.NX_WORKSPACE_ROOT_PATH = this.previousEnvWorkspaceRootPath;
+    setWorkspaceRoot(this.previousWorkspaceRoot);
   }
 
   reset() {
