+++ packages/nx/tsconfig.spec.json
@@ -12,7 +12,7 @@
     "**/*.spec.js",
     "**/*.spec.jsx",
     "**/*.d.ts",
-    "internal-testing-utils/**/*.ts",
+    "./src/internal-testing-utils/**/*.ts",
     "jest.config.ts"
   ]
 }
